backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic-settings==2.3.4
python-dotenv==1.0.1
requests==2.32.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pyyaml==6.0.2
tqdm==4.66.5
tabulate==0.9.0
numpy==1.26.4
